// tools/yaml-to-json-converter/client/types.ts
export type YamlToJsonConverterMessages = Readonly<{
  meta: {
    name: string
    description: string
  }
  importFromFileLabel: string
  downloadJsonLabel: string
  yamlLabel: string
  yamlDescription: string
  jsonLabel: string
  jsonDescription: string
  jsonEmptyDescription: string
  yamlPlaceholder: string
  invalidYamlLabel: string
  copyJsonLabel: string
  copiedLabel: string
}>